image_moderation.py

from typing import Dict, Any, List
from PIL import Image
import io

class ImageModerator:
    """Placeholder for image content moderation"""
    
    def __init__(self):
        self.enabled = False
    
    def analyze_image(self, image_data: bytes) -> Dict[str, Any]:
        """
        Analyze image for inappropriate content
        
        Returns:
            Dictionary with moderation results
        """
        if not self.enabled:
            return {
                "severity": 0.0,
                "detected_issues": [],
                "categories": {},
                "message": "Image moderation not enabled"
            }
        
        try:
            # Placeholder: In production, call actual image moderation API
            img = Image.open(io.BytesIO(image_data))
            
            # Mock analysis
            return {
                "severity": 0.0,
                "detected_issues": [],
                "categories": {
                    "adult": 0.0,
                    "violence": 0.0,
                    "drugs": 0.0,
                    "hate_symbols": 0.0
                },
                "image_info": {
                    "format": img.format,
                    "size": img.size,
                    "mode": img.mode
                },
                "message": "Image analyzed (placeholder)"
            }
        except Exception as e:
            return {
                "severity": 0.0,
                "detected_issues": ["error_processing_image"],
                "error": str(e)
            }
    
    def enable_moderation(self, api_key: str = None):
        """Enable image moderation with API credentials"""
        if api_key:
            self.enabled = True
            print("Image moderation enabled")
        else:
            print("No API key provided for image moderation")
