requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
redis==5.0.1
langgraph==0.0.20
langchain==0.1.0
langchain-anthropic==0.1.1
anthropic
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
langdetect==1.0.9
pillow==10.1.0
